analytics-service/requirements.txt
django
djangorestframework
clickhouse-driver
confluent-kafka
pydantic